+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - Fixed `mod.meta.current_path` and `mod.util.eval()` causing a panic when used in the XML Sandbox.
+- Improved error message for `extract` command failing extract an output file (now mentions path).
 
 ## [v0.7.2]